feat: make validation return side tables
